+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Added
+
+- Add inspection and quick fix for unused local variables.
+
 ## [1.5.0] - 2023-11-27
 
 ### Added